requirements.txt
pandas
openpyxl
google-adk
mcp
fastmcp
python-dotenv
litellm
fastapi
uvicorn
pydantic
mcp[cli]
google-generativeai
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
